fixed issue and some refactors
