+++ samples/Rules.Framework.WebUI.Sample/ReadmeExample/BasicRulesEngineExample.cs
@@ -1,7 +1,6 @@
 namespace Rules.Framework.WebUI.Sample.ReadmeExample
 {
     using System;
-    using global::Rules.Framework.Builder;
     using global::Rules.Framework.Core;
     using global::Rules.Framework.Providers.InMemory;
     using global::Rules.Framework.WebUI.Sample.Engine;
@@ -59,6 +58,8 @@ namespace Rules.Framework.WebUI.Sample.ReadmeExample
                 .WithName("Rule for September sample for premium clients.")
                 .WithContent(BasicContentType.FreeSample, "ConditionerSample")
                 .WithCondition(BasicConditionType.ClientType, Operators.Equal, "Premium")
+                .WithCondition(c => c
+                    .And(a => a.)
                 .WithDatesInterval(new DateTime(2023, 09, 01), new DateTime(2023, 10, 01))
                 .Build();
