+++ src/Rules.Framework/Builder/Deprecated/ConditionNodeBuilder.cs
