+++ src/Rules.Framework/Builder/Deprecated/IComposedConditionNodeBuilder.cs
