+++ src/Rules.Framework/Builder/Deprecated/IConditionNodeBuilder.cs
