+++ src/Rules.Framework/Builder/Deprecated/IValueConditionNodeBuilder.cs
